pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "nba-draft-ml"
version = "0.1.0"
description = "NBA Draft Prospect ML Research"
requires-python = ">=3.11,<3.13"

dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "transformers",
    "datasets",
    "pandas",
    "numpy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "pyyaml",
    "python-dotenv",
    "hydra-core",
    "tqdm",
    "wandb",
    "mlflow",
    "playwright",
    "nba-api>=1.11.4",
    "xgboost>=3.2.0",
    "polars>=1.40.1",
    "pyarrow>=23.0.1",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "ruff",
    "ipykernel",
    "jupyter",
    "notebook",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
